Remove pyftsubset, split into fontcull lib + fontcull-cli
- Remove pyftsubset fallback entirely, use klippa exclusively
- Split into two crates:
  - fontcull: Pure Rust font subsetting library (no browser deps)
  - fontcull-cli: CLI tool with browser-based glyph extraction
- klippa deps are now non-optional in the library
- Browser deps (chromiumoxide, tokio, etc.) only in CLI crate
- Update README to reflect new structure
- Add crates.io, docs.rs, and license badges
