+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
-    ".",
+    "fontcull",
+    "fontcull-cli",
     "vendored/fontcull-font-types",
     "vendored/fontcull-font-test-data",
     "vendored/fontcull-read-fonts",
@@ -8,59 +9,6 @@ members = [
     "vendored/fontcull-skrifa",
     "vendored/fontcull-klippa",
 ]
-
-[package]
-name = "fontcull"
-version = "1.0.5"
-edition = "2024"
-authors = ["Amos Wenger <amos@bearcove.eu>"]
-description = "Subset fonts based on actual glyph usage from web pages"
-license = "MIT"
-repository = "https://github.com/fasterthanlime/fontcull"
-keywords = ["font", "subset", "woff2", "web", "glyphhanger"]
-categories = ["command-line-utilities", "web-programming"]
-
-[[bin]]
-name = "fontcull"
-path = "src/main.rs"
-required-features = ["browser"]
-
-[lib]
-name = "fontcull"
-path = "src/lib.rs"
-
-[features]
-default = ["klippa", "browser"]
-klippa = ["dep:fontcull-klippa", "dep:fontcull-skrifa", "dep:fontcull-write-fonts", "dep:fontcull-read-fonts", "dep:woff"]
-static-analysis = ["dep:scraper"]
-browser = ["dep:chromiumoxide", "dep:tokio", "dep:serde_json", "dep:futures"]
-
-[dependencies]
-# CLI dependencies (always needed for binary)
-clap = { version = "4", features = ["derive"] }
-glob = "0.3"
-eyre = "0.6"
-color-eyre = "0.6"
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-serde = { version = "1", features = ["derive"] }
-url = "2"
-
-# Browser-based extraction (optional)
-chromiumoxide = { version = "0.7", features = ["tokio-runtime"], optional = true }
-tokio = { version = "1", features = ["full"], optional = true }
-serde_json = { version = "1", optional = true }
-futures = { version = "0.3", optional = true }
-
-# Static HTML/CSS analysis (optional)
-scraper = { version = "0.24", optional = true }
-
-# klippa backend (optional, vendored from googlefonts/fontations)
-fontcull-klippa = { version = "0.1.1", path = "vendored/fontcull-klippa", optional = true }
-fontcull-skrifa = { version = "0.39.1", path = "vendored/fontcull-skrifa", optional = true }
-fontcull-write-fonts = { version = "0.44.2", path = "vendored/fontcull-write-fonts", optional = true }
-fontcull-read-fonts = { version = "0.37.0", path = "vendored/fontcull-read-fonts", optional = true }
-woff = { version = "0.6", optional = true }
 
 # The profile that 'dist' will build with
 [profile.dist]
