+++ dist-workspace.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["cargo:."]
+members = ["cargo:fontcull-cli"]
 
 # Config for 'dist'
 [dist]
@@ -8,10 +8,14 @@ cargo-dist-version = "0.30.2"
 # CI backends to support
 ci = "github"
 # The installers to generate for each app
-installers = ["shell"]
+installers = ["shell", "powershell", "homebrew", "msi"]
 # Target platforms to build apps for (Rust target-triple syntax)
-targets = ["aarch64-apple-darwin", "aarch64-unknown-linux-gnu", "x86_64-apple-darwin", "x86_64-unknown-linux-gnu"]
+targets = ["aarch64-apple-darwin", "aarch64-unknown-linux-gnu", "x86_64-apple-darwin", "x86_64-unknown-linux-gnu", "x86_64-pc-windows-msvc"]
 # Path that installers should place binaries in
 install-path = "CARGO_HOME"
 # Whether to install an updater program
 install-updater = true
+# A GitHub repo to push Homebrew formulas to
+tap = "bearcove/homebrew-tap"
+# Publish jobs to run in CI
+publish-jobs = ["homebrew"]
