+++ fontcull-cli/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "fontcull-cli"
+version = "1.0.5"
+edition = "2024"
+authors = ["Amos Wenger <amos@bearcove.eu>"]
+description = "CLI tool to subset fonts based on actual glyph usage from web pages"
+license = "MIT"
+repository = "https://github.com/bearcove/fontcull"
+homepage = "https://github.com/bearcove/fontcull"
+keywords = ["font", "subset", "woff2", "web", "glyphhanger"]
+categories = ["command-line-utilities", "web-programming"]
+
+[package.metadata.wix]
+upgrade-guid = "9FF2FB57-9D1B-4336-BDE9-0A200F15DADA"
+path-guid = "D8A89A7F-FEAF-4CCB-BC7B-2B3D8D0E7093"
+license = false
+eula = false
+
+[[bin]]
+name = "fontcull"
+path = "src/main.rs"
+
+[dependencies]
+fontcull = { version = "1.0.5", path = "../fontcull" }
+
+# CLI dependencies
+clap = { version = "4", features = ["derive"] }
+glob = "0.3"
+color-eyre = "0.6"
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+
+# Browser-based extraction
+chromiumoxide = { version = "0.7", features = ["tokio-runtime"] }
+tokio = { version = "1", features = ["full"] }
+serde_json = "1"
+futures = "0.3"
