+++ fontcull/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "fontcull"
+version = "1.0.5"
+edition = "2024"
+authors = ["Amos Wenger <amos@bearcove.eu>"]
+description = "Pure Rust font subsetting library"
+license = "MIT"
+repository = "https://github.com/bearcove/fontcull"
+homepage = "https://github.com/bearcove/fontcull"
+keywords = ["font", "subset", "woff2", "ttf", "otf"]
+categories = ["text-processing", "web-programming"]
+
+[features]
+default = []
+static-analysis = ["dep:scraper"]
+
+[dependencies]
+# Static HTML/CSS analysis (optional)
+scraper = { version = "0.24", optional = true }
+
+# klippa backend (vendored from googlefonts/fontations)
+fontcull-klippa = { version = "0.1.1", path = "../vendored/fontcull-klippa" }
+fontcull-skrifa = { version = "0.39.1", path = "../vendored/fontcull-skrifa" }
+fontcull-write-fonts = { version = "0.44.2", path = "../vendored/fontcull-write-fonts" }
+fontcull-read-fonts = { version = "0.37.0", path = "../vendored/fontcull-read-fonts" }
+woofwoof = "1.0"
